rename: 홈페이지 스크롤 유도 텍스트 컴포넌트 파일명 변경
