+++ src/components/home/ScrollLeadText.tsx
